Fix more structs
