+++ src/panels/lovelace/editor/types.ts
@@ -1,18 +1,14 @@
 import {
-  any,
   array,
   boolean,
   dynamic,
   enums,
   literal,
-  number,
   object,
   optional,
   string,
-  type,
   union,
 } from "superstruct";
-import { custom } from "../../../common/structs/is-custom-type";
 import {
   ActionConfig,
   LovelaceCardConfig,
@@ -97,14 +93,10 @@ export interface EditSubElementEvent {
   subElementConfig: SubElementEditorConfig;
 }
 
-export const actionConfigStruct = dynamic((_value, ctx) => {
-  if (ctx.branch[0][ctx.path[0]]) {
-    return (
-      actionConfigMap[ctx.branch[0][ctx.path[0]].action] ||
-      actionConfigStructType
-    );
+export const actionConfigStruct = dynamic((value: any) => {
+  if (value && value.action in actionConfigMap) {
+    return actionConfigMap[value.action];
   }
-
   return actionConfigStructType;
 });
 
@@ -164,90 +156,6 @@ export const actionConfigStructType = object({
   confirmation: optional(actionConfigStructConfirmation),
 });
 
-const buttonEntitiesRowConfigStruct = object({
-  type: string(),
-  name: string(),
-  action_name: optional(string()),
-  tap_action: actionConfigStruct,
-  hold_action: optional(actionConfigStruct),
-  double_tap_action: optional(actionConfigStruct),
-});
-
-const castEntitiesRowConfigStruct = object({
-  type: string(),
-  view: union([string(), number()]),
-  dashboard: optional(string()),
-  name: optional(string()),
-  icon: optional(string()),
-  hide_if_unavailable: optional(boolean()),
-});
-
-const callServiceEntitiesRowConfigStruct = object({
-  type: string(),
-  name: string(),
-  service: string(),
-  icon: optional(string()),
-  action_name: optional(string()),
-  service_data: optional(any()),
-});
-
-const conditionalEntitiesRowConfigStruct = object({
-  type: string(),
-  row: any(),
-  conditions: array(
-    object({
-      entity: string(),
-      state: optional(string()),
-      state_not: optional(string()),
-    })
-  ),
-});
-
-const dividerEntitiesRowConfigStruct = object({
-  type: string(),
-  style: optional(any()),
-});
-
-const sectionEntitiesRowConfigStruct = object({
-  type: string(),
-  label: optional(string()),
-});
-
-const webLinkEntitiesRowConfigStruct = object({
-  type: string(),
-  url: string(),
-  name: optional(string()),
-  icon: optional(string()),
-});
-
-const buttonsEntitiesRowConfigStruct = object({
-  type: string(),
-  entities: array(
-    union([
-      object({
-        entity: string(),
-        icon: optional(string()),
-        image: optional(string()),
-        name: optional(string()),
-      }),
-      string(),
-    ])
-  ),
-});
-
-const attributeEntitiesRowConfigStruct = object({
-  type: string(),
-  entity: string(),
-  attribute: string(),
-  prefix: optional(string()),
-  suffix: optional(string()),
-  name: optional(string()),
-});
-
-const customRowConfigStruct = type({
-  type: custom(),
-});
-
 export const entitiesConfigStruct = union([
   object({
     entity: string(),
@@ -262,14 +170,4 @@ export const entitiesConfigStruct = union([
     double_tap_action: optional(actionConfigStruct),
   }),
   string(),
-  buttonEntitiesRowConfigStruct,
-  castEntitiesRowConfigStruct,
-  conditionalEntitiesRowConfigStruct,
-  dividerEntitiesRowConfigStruct,
-  sectionEntitiesRowConfigStruct,
-  webLinkEntitiesRowConfigStruct,
-  buttonsEntitiesRowConfigStruct,
-  attributeEntitiesRowConfigStruct,
-  callServiceEntitiesRowConfigStruct,
-  customRowConfigStruct,
 ]);
